More tests, new file structure.
Added some more tests for the new functions and created a new file,
sympy/combinatorics/testutil.py that will accomodate utility functions
used specifically in testing functions from the combinatorics module
